refactor: renaming for common component usage
